Add an "editmaps" UI for configuring destination maps
No need to fiddle with about:config anymore. Just open this panel from
the Add-on settings.

Features:
- modify (name, GeoHack template URI)
- enable/disable
- resort via HTML5 Drag & Drop of the "Name" column
- add new
- remove existing

Any change, even merely adding a letter to the name, is applied
immediately without having to press some OK button. There also is no
cancel or undo button.
Incomplete entries (only consisting of the name or the URL) will be
discarded on panel closing.

Changes in about:config are still possible and do not interfere with
each other.

The object array destmaps is hidden from simple prefs now, it is not
needed anymore.