usemaps: prevent visible panel resize and content composition
Especially this was noticeable for animated resize.

This also circumvents a bug in GTK+ 3.20.2, which led to wrong panel position
from the second panel activation on. The panel was shifted to the left so that
the upper left corner was placed at the same position as before the resize.
Fixed in GTK+ 3.20.3, see also:
    https://bugs.archlinux.org/task/48853
    https://bugzilla.gnome.org/show_bug.cgi?id=764730