geourl: support for GeoHack URLs
There are several formats available for defining latitude/longitude. Only the
ones using N/S and E/W are supported, the format "lat;lon" as well as the "nice
URLs" are unsupported for now. But they don't seem to be used in Wikipedia as
the main GeoHack user.

Also extend tests for calculation from {scale} to {osmzoom} for the scale
values based on the "_type" URL parameter.