Detect URL change in address bar for all tabs
Before this change, it was only possible on the first tab for each window, now
it really works.

Furthermore, duplicated events (two instead of one) on URL change do not occur
now anymore.

In this change it didn't make much difference whether addProgressListener() or
addTabsProgressListener() has been used, but the latter avoids an additional
event on creation of a tab.

To reproduce this misbehavior:
- scrolling on https://openinframap.org/#2/46.992/8.552
  It switches between integer and (currently unsupported) float numbers
  for zoom in new style OSM URL.
- moving the map on e.g. https://www.openstreetmap.org/changeset/1
  It adds the #map= coordinates.