Add a context menu for use on hyperlinks of supported Geo URLs
A hyperlink <a href> with a supported Geo URL can now be opened directly
to the destionation map of your choice. The items of the context menu
are the same as configured for the button panel.

The context menu is always visible when clicking on a link. If the link
is not supported, the disabled icon is used as indicator in the context
menu instead of the colored one. Not just hiding the context menu has
technical reasons - this might be changed in the future.