Let the browser handle the disabled appearance of the toolbar icon
If the toggle button is in disabled state, the icon is automatically displayed
inactive by the browser (maybe this was not the case in older Firefox
versions?). Hence using a gray icon led to very light gray.

Now the icon is also colored in disabled state, but with pale appearance.