Fix regression with undesired disabled globe button
If the usemaps panel was visible and you did not click on a map item,
but somewhere else, the panel was hidden as usual, but the globe button
was disabled although the tab still contained the same supported map.
This wasn't caused by setShowmapButtonState(), this function was not
invoked. For some reason the global shomwapButton.disabled state was
used (the bad behavior didn't occur when changing it to true in
setShowmapButtonState()).

Setting the disabled state tab- instead of window-specific fixes this
issue.

Regression introduced in commit af45e76 (Enable/disable the globe button
separately for each browser window, 2018-12-21).