Add a "geosite" framework to use geo positions in websites with the usemaps panel
This new functionality will be used for websites whose URL does not contain geo
information or is not updated during map dragging. This also includes websites
not containing a map at all, but other content corresponding to a geo position
such as geo referenced photos.

No website scanners implemented yet, but prepared for simple addition.

The return type of scanGeosite() is an object currently only containing
GeoHackCoords as property. Thus it is extensible for the csae that the content
script (scanner) does not determine the coordinates itself, but e.g. finds a
map URL, which can then be passed to the Add-On for use with the "geourl" lib.

The message from content script to Add-On script has been added to the
"contentScript" property instead of invoking it in the geosite.js content
script, so it is kept clean from Add-On API code.