geosite: support for "geo tag" meta elements
Unsure about how widespread these tags are used, at least it works for the
photo platform "locr".

Because the globe button is disabled first and later (after receiving the
message from content script) enabled, there is a little flicker in the icon
each time when (re-)loading the site or when switching to the tab from
another tab where the button was enabled. In the former case even twice
because of two different tab events occur for updating the button state.

Unfortunately, using the page-worker API for regression tests did not work, but
led to this error:

  JPM [error]   Message: TypeError: remoteFrame is null

So before adding tests for this geotag and other geosite scanners, the test
framework will be changed to something independent on JPM. This will be
necessary when switching to WebExtensions anyway.