geosite: support for Wikipedia
Articles with a geo position can now be used for showing the map of your
choice.

Unfortunately the Wikipedia sites are not implemented identically across
different languages. Beside different DOM structure, there are different
element ID or CSS class names in use. Thus for searching the GeoHack
link, no such names are used now except the direct anchor class names.

The Kartographer link also seems to be used consistently only in the
anchor itself, but not in the parent elements. Checked with French,
Russian and Portuguese.

In the Russian site, both the GeoHack and the Kartographer link is included,
but the GeoHack has priority.