Fix Firefox version handling if extensions.lastAppVersion is not set
It was already intended for the fallback to assume being the latest version and
having all features possible. So the return condition in firefoxCanDimIcon()
should have been

  return firefoxVersion == 0 || firefoxVersion >= 55;

with the fallback set to 0. But better use a fallback of 56, which is the last
version of Firefox with legacy Add-On support. Even if Waterfox 57 should
support them, probably there won't be new features added for the need to
distinguish further versions.

But the scope of support for this legacy Add-On is version 52 (ESR) to 56
anyway.