geosite: support for Flickr photo site
Does not work reliably:

1) Sometimes in Flickr's photostream feature (browsing through photos
with left/right buttons) on a photo site WITHOUT geo information, the
toggle button is enabled, but nothing happens when clicking.
Reason: On the first location change event, the Flickr map link from the
previous photo site WITH geo information is found, on the second event,
the DOM is updated and null is returned. Because of the asynchronous
nature of the content scripts, both results arrive after the two geourl
address bar checks (which are false), so the button is enabled by the
first wrong result, but not disabled by the second address bar check.
The second and correct result does not enable the button because this
race condition is not considered yet.
This behavior seems to occur always when switching from a photo site
WITH to a site WITHOUT geo information.

2) Sometimes on a photo site WITH geo information, the toggle button is
disabled. The DOM does not seem to be updated before the last event.