Use AVA for unit tests instead of jpm
Preparation for switching to WebExtensions, but replacing jpm has other
benefits like faster execution and nicer output.

One geourl parse test had to be modified to avoid an error due to
duplicated test names.

The UI tests (only 1 existing) have been removed. It is not easily
possible to convert it due to its SDK dependencies. With WebExtensions
it would have to be adapted anyway.